Unescaping texts in db, changing views.
Small but scary chunk of code. Made migration myself, changed data
throughout database, then changed methods of creating, updating and
viewing texts. Escaping moved from views to templates, where it probably belongs.

The idea of changing likes from intermediate tables to django many-to-many
relationships was deprecated after a discussion with the teacher; to
create dislikes we need 'status' field in that table, and for that we
obviously need table. Will be in the next commit, along with
'categories'. In my implementation - tag's analogue.