Hidden comments' fields.
Oh, that was painful. Probably, the most painful part for now. My hatred
to django's forms, class-based views and different little attributes
that either don't exist at all or are blank because of moon's phase
during the creation of that particular piece of multiply inherited
$%!* of code.

Not that writing your first js script in jQuery and ajax is wonderful
and funny. Bloodbath is somewhat better description. But, well, there's
at least a little hope of understanding an underlying logic. Not the
case for django. Because, you understand, there's one --
and only one -- obvious way to do it. I'm just not Dutch.